Replaces merit with position for ranking
Updates the rankings model to use a 'position' attribute instead of 'merit'.
This simplifies ranking logic and aligns with the intended use case of ordering teams.

Removes unused methods and validations related to the previous 'merit' system.

Introduces a `find_or_create_rankings` method to the User model. This method
ensures that a user has a ranking for each available team within a given
category, creating them if they don't already exist.
This sets a default ordering which can then be modified by a user.