Adds data import rake task and improves result calculation
Adds a rake task to import team and user data from a CSV file.
This task streamlines the initial setup and population of the database.
It handles team creation/update and user creation/update (as team members).

Improves the result calculation logic in the ResultsController.
This change refactors the calculation to improve efficiency and
correctly handle cases where only team member or judge/admin rankings are available.

Adds the 'csv' gem for parsing the import file.
Adds the imported file to the .gitignore.