Adds system tests for teams and user factories
Adds system tests for creating teams and navigating to the
teams index page.

Adds an after create hook to the user factory to nullify invitation
tokens.

Removes scaffold generated request and view tests, and routing specs
for rankings, teams and users as these are not used.

Updates TODO.md with new tasks.